# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "NestedWorkoutItemRequest": {
        "properties": {
          "duration_minutes": {
            "default": 0,
            "minimum": 0.0,
            "title": "Duration Minutes",
            "type": "number"
          },
          "exercise_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Exercise Name",
            "type": "string"
          },
          "id": {
            "default": 0,
            "minimum": 0.0,
            "title": "Id",
            "type": "integer"
          },
          "reps": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Reps",
            "type": "integer"
          },
          "sets": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Sets",
            "type": "integer"
          },
          "weight": {
            "default": 0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          },
          "workout_id": {
            "default": 0,
            "minimum": 0.0,
            "title": "Workout Id",
            "type": "integer"
          }
        },
        "required": [
          "exercise_name"
        ],
        "title": "NestedWorkoutItemRequest",
        "type": "object"
      },
...
      "WorkoutItemRequest": {
        "properties": {
          "duration_minutes": {
            "default": 0,
            "minimum": 0.0,
            "title": "Duration Minutes",
            "type": "number"
          },
          "exercise_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Exercise Name",
            "type": "string"
          },
          "reps": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Reps",
            "type": "integer"
          },
          "sets": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Sets",
            "type": "integer"
          },
          "weight": {
            "default": 0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "exercise_name"
        ],
        "title": "WorkoutItemRequest",
        "type": "object"
      },
      "WorkoutItemResponse": {
        "properties": {
          "duration_minutes": {
            "default": 0,
            "title": "Duration Minutes",
            "type": "number"
          },
          "exercise_name": {
            "title": "Exercise Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "reps": {
            "default": 0,
            "title": "Reps",
            "type": "integer"
          },
          "sets": {
            "default": 0,
            "title": "Sets",
            "type": "integer"
          },
          "weight": {
            "default": 0,
            "title": "Weight",
            "type": "number"
          },
          "workout_id": {
            "title": "Workout Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "workout_id",
          "exercise_name",
          "sets",
          "reps",
          "weight",
          "duration_minutes"
        ],
        "title": "WorkoutItemResponse",
        "type": "object"
      },
      "WorkoutRequest": {
        "properties": {
          "duration": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Duration",
            "type": "integer"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NestedWorkoutItemRequest"
                },
                "maxItems": 500,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 8601 with offset; null means unknown; omitted on PUT preserves",
            "title": "Occurred At"
          },
          "workout_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Workout Name",
            "type": "string"
          }
        },
        "required": [
          "workout_name"
        ],
        "title": "WorkoutRequest",
        "type": "object"
      },
      "WorkoutResponse": {
        "properties": {
          "duration": {
            "title": "Duration",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkoutItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "workout_name": {
            "title": "Workout Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "workout_name",
          "duration",
          "occurred_at"
        ],
        "title": "WorkoutResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "MetricsBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "GopherFit API",
    "version": "2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/workouts/": {
      "get": {
        "operationId": "workouts_workouts__get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkoutResponse"
                  },
                  "title": "Response Workouts Workouts  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Workouts",
        "tags": [
          "workouts"
        ]
      },
      "post": {
        "operationId": "create_workout_workouts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Workout",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/{id}": {
      "delete": {
        "operationId": "delete_workout_workouts__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Workout",
        "tags": [
          "workouts"
        ]
      },
      "get": {
        "operationId": "workout_workouts__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "ge": -9223372036854775808,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Workout",
        "tags": [
          "workouts"
        ]
      },
      "put": {
        "operationId": "update_workout_workouts__id__put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "ge": -9223372036854775808,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Workout",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/{id}/items": {
      "post": {
        "operationId": "create_item_workouts__id__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Item",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/{id}/items/{itemId}": {
      "delete": {
        "operationId": "delete_item_workouts__id__items__itemId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "Itemid",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Item",
        "tags": [
          "workouts"
        ]
      },
      "put": {
        "operationId": "update_item_workouts__id__items__itemId__put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "Itemid",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Item",
        "tags": [
          "workouts"
        ]
      }
    }
  }
}